Refactor common methods into TaskAPI module
Developers expect a consistent API writing tasks and helpers. Before,
this consistency was implemented through copy-and-paste. Now it is
better enforced via a `TaskAPI` module that is included in both `Remote`
and `TaskLibrary`.
